@@ -10,7 +10,7 @@
     "build:docker": "yarn build:dist && docker build -t secret-agent .",
     "build:ci": "yarn tsc && cd build && yarn install",
     "copy:node_modules": "shx cp -r node_modules \"build/node_modules\"",
-    "copy:build": "copyfiles \"testing/*/**\" \"puppet/test/*/**\" \"yarn.lock\" \"mitm-socket/dist/*\" -a build && shx cp package.build.json build/package.json",
+    "copy:build": "copyfiles \"testing/*/**\" \"**/*.mjs\" \"puppet/test/*/**\" \"yarn.lock\" \"mitm-socket/dist/*\" -a build && shx cp package.build.json build/package.json",
     "copy:dist": "copyfiles \"mitm-socket/lib/*\" -e \"mitm-socket/dist\" \"**/.gitignore\" \"**/.npmignore\" build-dist",
     "tsc": "tsc -b tsconfig.json && yarn copy:build && node prepare-build.js && yarn workspace @secret-agent/replay build:backend",
     "watch": "tsc-watch -b -w tsconfig.json --onSuccess \"yarn workspace @secret-agent/replay build:backend-paths\"",
@@ -41,38 +41,38 @@
     }
   },
   "devDependencies": {
-    "@commitlint/cli": "^9.1.1",
-    "@commitlint/config-conventional": "^9.1.1",
-    "@types/jest": "^26.0.10",
-    "@types/node": "^12.7.11",
-    "@typescript-eslint/eslint-plugin": "^3.10.1",
-    "@typescript-eslint/parser": "^3.10.1",
-    "copyfiles": "^2.3.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.7.0",
-    "eslint-config-airbnb-typescript": "^9.0.0",
-    "eslint-config-prettier": "^6.11.0",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^9.1.2",
+    "@types/jest": "^26.0.16",
+    "@types/node": "^13.13.34",
+    "@typescript-eslint/eslint-plugin": "^4.9.0",
+    "@typescript-eslint/parser": "^4.9.0",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.14.0",
+    "eslint-config-airbnb-typescript": "^12.0.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-monorepo-cop": "^1.0.2",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^23.20.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-promise": "^4.2.1",
-    "husky": "^4.2.5",
-    "jest": "^26.4.2",
-    "jest-environment-node": "^26.3.0",
+    "husky": "^4.3.0",
+    "jest": "^26.6.3",
+    "jest-environment-node": "^26.6.2",
     "jest-summary-reporter": "^0.0.2",
     "lerna": "^3.22.1",
-    "lint-staged": "^10.2.11",
-    "prettier": "^1.19.1",
-    "pretty-quick": "^2.0.1",
-    "shx": "^0.3.2",
-    "source-map-support": "^0.5.16",
+    "lint-staged": "^10.5.2",
+    "prettier": "^2.2.1",
+    "pretty-quick": "^3.1.0",
+    "shx": "^0.3.3",
+    "source-map-support": "^0.5.19",
     "tsc-watch": "^4.2.9",
-    "typescript": "~3.8.3"
+    "typescript": "~4.1.2"
   },
   "workspaces": {
     "packages": [
