@@ -11,9 +11,9 @@
   },
   "dependencies": {
     "@babel/runtime-corejs3": "^7.11.2",
-    "@vue/cli-plugin-babel": "~4.5.4",
-    "@vue/cli-plugin-typescript": "~4.5.4",
-    "@vue/cli-service": "~4.5.4",
+    "@vue/cli-plugin-babel": "~4.5.9",
+    "@vue/cli-plugin-typescript": "~4.5.9",
+    "@vue/cli-service": "~4.5.9",
     "animejs": "^3.2.0",
     "core-js": "^3.6.5",
     "source-map-support": "^0.5.19",
@@ -32,7 +32,7 @@
     "pug": "^2.0.4",
     "pug-plain-loader": "^1.0.0",
     "sass-loader": "^8.0.2",
-    "typescript": "~3.8.3",
+    "typescript": "~4.1.2",
     "webpack": "4.44.2"
   }
 }
