@@ -66,17 +66,16 @@
   },
   "devDependencies": {
     "@types/chrome": "^0.0.113",
-    "@types/electron": "^1.6.10",
-    "@types/node": "^12.7.11",
+    "@types/node": "^13.13.34",
     "concurrently": "^5.2.0",
     "core-js": "^3.6.5",
-    "cross-env": "^7.0.2",
+    "cross-env": "^7.0.3",
     "devtron": "^1.4.0",
     "electron": "10.1.5",
     "electron-builder": "^22.9.1",
     "electron-icon-builder": "^1.0.2",
-    "shx": "^0.3.2",
+    "shx": "^0.3.3",
     "tscpaths": "^0.0.9",
-    "typescript": "~3.8.3"
+    "typescript": "~4.1.2"
   }
 }
