@@ -41,7 +41,7 @@
     "sass-loader": "^8.0.0",
     "sharp": "^0.23.2",
     "ts-loader": "^6.2.1",
-    "typescript": "~3.8.3",
+    "typescript": "~4.1.2",
     "vue-svg-loader": "^0.12.0",
     "vue-template-compiler": "^2.6.12"
   },
