@@ -1,7 +1,4 @@
-const SecretAgent = require('@secret-agent/full-client');
-
-process.env.SHOW_BROWSER = 'true';
-process.env.SA_REPLAY_DEBUG = '1';
+import SecretAgent from '@secret-agent/full-client';
 
 (async () => {
   const agent = await new SecretAgent();
@@ -9,9 +6,9 @@ process.env.SA_REPLAY_DEBUG = '1';
   const url = `https://dataliberationfoundation.org/`;
   console.log('Opened Browser');
 
-  await browser.goto(url);
-  await browser.waitForAllContentLoaded();
+  await agent.goto(url);
+  await agent.waitForAllContentLoaded();
 
-  await browser.waitForMillis(5e3);
-  await browser.close();
+  await agent.waitForMillis(5e3);
+  await agent.close();
 })();
